# compute: add notebook block component