--- a/client/web/src/notebooks/blocks/compute/NotebookComputeBlock.tsx
+++ b/client/web/src/notebooks/blocks/compute/NotebookComputeBlock.tsx
@@ -1,5 +1,6 @@
 import classNames from 'classnames'
 import React, { useRef } from 'react'
+import ElmComponent from 'react-elm-components'
 
 import { ThemeProps } from '@sourcegraph/shared/src/theme'
 
@@ -10,10 +11,66 @@ import blockStyles from '../NotebookBlock.module.scss'
 import { useBlockSelection } from '../useBlockSelection'
 import { useBlockShortcuts } from '../useBlockShortcuts'
 
+import { Elm } from './component/src/Main.elm'
 import styles from './NotebookComputeBlock.module.scss'
 
 interface ComputeBlockProps extends BlockProps, ComputeBlock, ThemeProps {
     isMacPlatform: boolean
+}
+
+interface ElmEvent {
+    data: string
+    eventType?: string
+    id?: string
+}
+
+function setupPorts(ports: {
+    receiveEvent: { send: (event: ElmEvent) => void }
+    openStream: { subscribe: (callback: (args: string[]) => void) => void }
+}): void {
+    const sources: { [key: string]: EventSource } = {}
+
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    function sendEventToElm(event: any): void {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
+        const elmEvent = { data: event.data, eventType: event.type || null, id: event.id || null }
+        ports.receiveEvent.send(elmEvent)
+    }
+
+    function newEventSource(address: string): EventSource {
+        sources[address] = new EventSource(address)
+        return sources[address]
+    }
+
+    function deleteAllEventSources(): void {
+        for (const [key] of Object.entries(sources)) {
+            deleteEventSource(key)
+        }
+    }
+
+    function deleteEventSource(address: string): void {
+        sources[address].close()
+        delete sources[address]
+    }
+
+    ports.openStream.subscribe((args: string[]) => {
+        deleteAllEventSources() // Close any open streams if we receive a request to open a new stream before seeing 'done'.
+        console.log(`stream: ${args[0]}`)
+        const address = args[0]
+
+        const eventSource = newEventSource(address)
+        eventSource.addEventListener('error', () => {
+            console.log('EventSource failed')
+        })
+        eventSource.addEventListener('results', sendEventToElm)
+        eventSource.addEventListener('alert', sendEventToElm)
+        eventSource.addEventListener('error', sendEventToElm)
+        eventSource.addEventListener('done', () => {
+            deleteEventSource(address)
+            // Note: 'done:true' is sent in progress too. But we want a 'done' for the entire stream in case we don't see it.
+            sendEventToElm({ type: 'done', data: '' })
+        })
+    })
 }
 
 export const NotebookComputeBlock: React.FunctionComponent<ComputeBlockProps> = ({
@@ -78,7 +135,10 @@ export const NotebookComputeBlock: React.FunctionComponent<ComputeBlockProps> = 
                 aria-label="Notebook compute block"
                 ref={blockElement}
             >
-                <div className="elm" />
+                <div className="elm">
+                    {/* eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access */}
+                    <ElmComponent src={Elm.Main} ports={setupPorts} flags={null} />
+                </div>
             </div>
             {blockMenu}
         </div>