--- a/client/web/src/notebooks/blocks/compute/component/elm.json
+++ b/client/web/src/notebooks/blocks/compute/component/elm.json
@@ -1,0 +1,34 @@
+{
+  "type": "application",
+  "source-directories": ["src"],
+  "elm-version": "0.19.1",
+  "dependencies": {
+    "direct": {
+      "NoRedInk/elm-json-decode-pipeline": "1.0.0",
+      "elm/browser": "1.0.2",
+      "elm/core": "1.0.5",
+      "elm/html": "1.0.0",
+      "elm/json": "1.1.3",
+      "elm/svg": "1.0.1",
+      "elm/time": "1.0.0",
+      "elm/url": "1.0.0",
+      "mdgriffith/elm-ui": "1.1.8",
+      "terezka/elm-charts": "3.0.0"
+    },
+    "indirect": {
+      "danhandrea/elm-time-extra": "1.1.0",
+      "debois/elm-dom": "1.3.0",
+      "elm/parser": "1.1.0",
+      "elm/virtual-dom": "1.0.2",
+      "justinmimbs/date": "3.2.1",
+      "justinmimbs/time-extra": "1.1.0",
+      "myrho/elm-round": "1.0.4",
+      "ryannhg/date-format": "2.3.0",
+      "terezka/intervals": "2.0.1"
+    }
+  },
+  "test-dependencies": {
+    "direct": {},
+    "indirect": {}
+  }
+}