--- a/client/web/src/notebooks/blocks/compute/elm.d.ts
+++ b/client/web/src/notebooks/blocks/compute/elm.d.ts
@@ -1,0 +1,5 @@
+declare module '*.elm' {
+    export const Elm: any
+}
+
+declare module 'react-elm-components'